+++ CHANGELOG.md
@@ -8,6 +8,8 @@
   (for [#3466](https://github.com/pbek/QOwnNotes/issues/3466))
 - Fixed extra top/right padding in the QLiteHtml preview panel layout
   (for [#3466](https://github.com/pbek/QOwnNotes/issues/3466))
+- Fixed QLiteHtml preview text selection to allow click-drag and shift-click selection
+  (for [#3466](https://github.com/pbek/QOwnNotes/issues/3466))
 
 ## 26.2.7